# ARROW-10810: [Rust] Speed up comparison kernels